Move more logic to wire test base class
