+++ seed/python-sdk/imdb/tests/custom/imdb/test_client.py
@@ -6,7 +6,6 @@ We verify wire compatibility against the API specification by:
     3. Returning a mocked response and verifying the client's handling.
 """
 
-import difflib
 import json
 import pytest
 
@@ -15,27 +14,6 @@ from seed.core.api_error import ApiError
 
 from .mock_server import MockResponse
 from .wire_test_base import WireTestBase
-
-
-# TODO(rmehndiratta): Move this to a test utils package
-def assert_json_eq(expected: str, actual: str):
-    expected = json.loads(expected)
-    actual = json.loads(actual)
-
-    expected = json.dumps(expected, indent=2, sort_keys=True)
-    actual = json.dumps(actual, indent=2, sort_keys=True)
-
-    if expected != actual:
-        diff = "\n".join(
-            difflib.unified_diff(
-                expected.splitlines(),
-                actual.splitlines(),
-                fromfile="expected",
-                tofile="actual",
-                lineterm="",
-            )
-        )
-        raise AssertionError(f"JSON mismatch:\n\n{diff}")
 
 
 class TestGetMovie(WireTestBase):
@@ -57,7 +35,7 @@ class TestGetMovie(WireTestBase):
         expected_response = json.dumps(movie_data)
         actual_response = movie.model_dump_json()
 
-        assert_json_eq(expected_response, actual_response)
+        self.assert_json_eq(expected_response, actual_response)
 
     def test_get_movie_not_found(self):
         movie_id = "tt0000000"
@@ -74,6 +52,7 @@ class TestGetMovie(WireTestBase):
         with pytest.raises(MovieDoesNotExistError) as exc_info:
             self.client.imdb.get_movie(movie_id=movie_id)
         assert exc_info.value.status_code == 404
+
         assert exc_info.value.body == movie_id
 
     def test_get_movie_server_error(self):
@@ -92,7 +71,11 @@ class TestGetMovie(WireTestBase):
         with pytest.raises(ApiError) as exc_info:
             self.client.imdb.get_movie(movie_id=movie_id)
         assert exc_info.value.status_code == 500
-        assert exc_info.value.body == error_response
+
+        expected_response = json.dumps(error_response)
+        actual_response = json.dumps(exc_info.value.body)
+
+        self.assert_json_eq(expected_response, actual_response)
 
 
 # TODO: add TestCreateMovie tests
